// src/lib/data/stats.generated.ts
export interface RealAgg {
  fights: number; minutes: number;
  sigL: number; sigA: number; oSigL: number; oSigA: number;
  tdL: number; tdA: number; oTdL: number; oTdA: number;
  kd: number; oKd: number; sub: number; ctrl: number;
}